macOS: move MenuShortcutManager to a separate file
